refactor: extract group creation from UI to createGroupFromTemplate util
