+++ src/modules/diet/template/application/templateToItem.ts
@@ -1,7 +1,7 @@
 import { calcRecipeMacros } from '~/legacy/utils/macroMath'
 import { type Item } from '~/modules/diet/item/domain/item'
-import { type Recipe } from '~/modules/diet/recipe/domain/recipe'
 import { type RecipeItem } from '~/modules/diet/recipe-item/domain/recipeItem'
+import { type Recipe } from '~/modules/diet/recipe/domain/recipe'
 import {
   type Template,
   isTemplateFood,
