+++ src/sections/search/components/ExternalTemplateToItemGroupModal.tsx
@@ -1,20 +1,14 @@
 import { type Accessor, type Setter } from 'solid-js'
-import {
-  type ItemGroup,
-  type RecipedItemGroup,
-  type SimpleItemGroup,
-  createRecipedItemGroup,
-  createSimpleItemGroup,
-} from '~/modules/diet/item-group/domain/itemGroup'
-import { type Recipe } from '~/modules/diet/recipe/domain/recipe'
+import { type ItemGroup } from '~/modules/diet/item-group/domain/itemGroup'
 import { type TemplateItem } from '~/modules/diet/template-item/domain/templateItem'
+import { createGroupFromTemplate } from '~/modules/diet/template/application/createGroupFromTemplate'
+import { templateToItem } from '~/modules/diet/template/application/templateToItem'
+import { type Template } from '~/modules/diet/template/domain/template'
 import { showError } from '~/modules/toast/application/toastManager'
 import { ModalContextProvider } from '~/sections/common/context/ModalContext'
 import { ItemEditModal } from '~/sections/food-item/components/ItemEditModal'
 import { handleApiError } from '~/shared/error/errorHandler'
 import { formatError } from '~/shared/formatError'
-import { templateToItem } from '~/modules/diet/template/application/templateToItem'
-import { type Template } from '~/modules/diet/template/domain/template'
 
 export type ExternalTemplateToItemGroupModalProps = {
   visible: Accessor<boolean>
@@ -39,34 +33,20 @@ export function ExternalTemplateToItemGroupModal(
           enable: true,
         })}
         onApply={(item) => {
-          if (item.__type === 'Item') {
-            const newGroup: SimpleItemGroup = createSimpleItemGroup({
-              name: item.name,
-              items: [item],
+          const template = props.selectedTemplate()
+          const { newGroup, operation, templateType } = createGroupFromTemplate(
+            template,
+            item,
+          )
+
+          props.onNewItemGroup(newGroup, item).catch((err) => {
+            handleApiError(err, {
+              component: 'ExternalTemplateToItemGroupModal',
+              operation,
+              additionalData: { itemName: item.name, templateType },
             })
-            props.onNewItemGroup(newGroup, item).catch((err) => {
-              handleApiError(err, {
-                component: 'ExternalTemplateToItemGroupModal',
-                operation: 'addSimpleItem',
-                additionalData: { itemName: item.name, templateType: 'Item' },
-              })
-              showError(`Erro ao adicionar item: ${formatError(err)}`)
-            })
-          } else {
-            const newGroup: RecipedItemGroup = createRecipedItemGroup({
-              name: item.name,
-              recipe: (props.selectedTemplate() as Recipe).id,
-              items: [...(props.selectedTemplate() as Recipe).items],
-            })
-            props.onNewItemGroup(newGroup, item).catch((err) => {
-              handleApiError(err, {
-                component: 'ExternalTemplateToItemGroupModal',
-                operation: 'addRecipeItem',
-                additionalData: { itemName: item.name, templateType: 'Recipe' },
-              })
-              showError(`Erro ao adicionar item: ${formatError(err)}`)
-            })
-          }
+            showError(`Erro ao adicionar item: ${formatError(err)}`)
+          })
         }}
       />
     </ModalContextProvider>
